Release candidate - ynEigen with ynLSDe Release

+++ src/external/tokens/IWETH.sol
@@ -3,10 +3,10 @@ pragma solidity ^0.8.24;
 
 interface IWETH {
     function deposit() external payable;
-    function withdraw(uint wad) external;
-    function totalSupply() external view returns (uint);
-    function approve(address guy, uint wad) external returns (bool);
-    function transfer(address dst, uint wad) external returns (bool);
-    function transferFrom(address src, address dst, uint wad) external returns (bool);
+    function withdraw(uint256 wad) external;
+    function totalSupply() external view returns (uint256);
+    function approve(address guy, uint256 wad) external returns (bool);
+    function transfer(address dst, uint256 wad) external returns (bool);
+    function transferFrom(address src, address dst, uint256 wad) external returns (bool);
     function symbol() external returns (string memory);
 }
